Added a WindowFactory to enable to pipe the window events to an object (in this case Engine.cpp).
A first folder structure has been added to the project.

Now the StartUp class only has the main WinAPI enttrance function. It creates the engine and starts it. The engine is running in an infitnite loop (Engine.run()), waiting to window events. When the closing event appears, the program is completly destroyed.